backend/requirements.txt
fastapi
uvicorn
python-docx
beautifulsoup4
mammoth
pdf2docx
html2docx
weasyprint
pydantic
requests
